+++ lib/shopify_cli/theme/dev_server/hot_reload.rb
@@ -14,8 +14,8 @@ module ShopifyCLI
           @theme = theme
           @mode = mode
           @streams = SSE::Streams.new
-          @remote_file_reloader = RemoteFileReloader.new(ctx, theme: @theme, streams: @streams)
-          @remote_file_deleter = RemoteFileDeleter.new(ctx, theme: @theme, streams: @streams)
+          # @remote_file_reloader = RemoteFileReloader.new(ctx, theme: @theme, streams: @streams)
+          # @remote_file_deleter = RemoteFileDeleter.new(ctx, theme: @theme, streams: @streams)
           @sections_index = SectionsIndex.new(@theme)
           @watcher = watcher
           @watcher.add_observer(self, :notify_streams_of_file_change)
@@ -41,69 +41,31 @@ module ShopifyCLI
         end
 
         def notify_streams_of_file_change(modified, added, removed)
-          files = (modified + added)
-            .reject { |file| @ignore_filter&.ignore?(file) }
-            .map { |file| @extension_mode ? @extension[file] : @theme[file] }
-
-          files -= liquid_css_files = files.select(&:liquid_css?)
-
-          deleted_files = removed
-            .reject { |file| @ignore_filter&.ignore?(file) }
-            .map { |file| @theme[file] }
-
-          remote_delete(deleted_files) unless deleted_files.empty?
-          reload_page(removed) unless deleted_files.empty?
-
-          hot_reload(files) unless files.empty?
-          remote_reload(liquid_css_files)
+          # TODO: call hooks here
         end
 
         private
-
-        def hot_reload(files)
-          paths = files.map(&:relative_path)
-          @streams.broadcast(JSON.generate(modified: paths))
-          @ctx.debug("[HotReload] Modified #{paths.join(", ")}")
-        end
-
-        def reload_page(removed)
-          @streams.broadcast(JSON.generate(reload_page: true))
-          @ctx.debug("[ReloadPage] Deleted #{removed.join(", ")}")
-        end
-
-        def remote_delete(files)
-          files.each do |file|
-            @ctx.debug("delete file each -> file.relative_path #{file.relative_path}")
-            @remote_file_deleter.delete(file)
-          end
-        end
-
-        def remote_reload(files)
-          files.each do |file|
-            @ctx.debug("reload file each -> file.relative_path #{file.relative_path}")
-            @remote_file_reloader.reload(file)
-          end
-        end
 
         def request_is_html?(headers)
           headers["content-type"]&.start_with?("text/html")
         end
 
         def inject_hot_reload_javascript(body)
-          hot_reload_filename = @extension_mode ? "hot-reload-tae.js" : "hot-reload-theme.js"
-          hot_reload_no_script = ::File.read("#{__dir__}/hot-reload-no-script.html")
-          hot_reload_js = ::File.read("#{__dir__}/#{hot_reload_filename}")
-          hot_reload_script = [
-            hot_reload_no_script,
-            "<script>",
-            params_js,
-            hot_reload_js,
-            "</script>",
-          ].join("\n")
-
-          body = body.join.gsub("</body>", "#{hot_reload_script}\n</body>")
-
-          [body]
+          # hot_reload_filename = @extension_mode ? "hot-reload-tae.js" : "hot-reload-theme.js"
+          # hot_reload_no_script = ::File.read("#{__dir__}/hot-reload-no-script.html")
+          # hot_reload_js = ::File.read("#{__dir__}/#{hot_reload_filename}")
+          # hot_reload_script = [
+          #   hot_reload_no_script,
+          #   "<script>",
+          #   params_js,
+          #   hot_reload_js,
+          #   "</script>",
+          # ].join("\n")
+          #
+          # body = body.join.gsub("</body>", "#{hot_reload_script}\n</body>")
+          #
+          # [body]
+          # TODO: call JS hook here
         end
 
         def params_js
