+++ lib/shopify_cli/theme/dev_server/hot_reload/hook.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module ShopifyCLI
+  module Theme
+    module DevServer
+      class HotReload
+        class Hook
+          def call
+            raise "#{self.class.name}#call must be defined!"
+          end
+        end
+      end
+    end
+  end
+end
