+++ lib/shopify_cli/theme/dev_server/hot_reload/hooks/reload_js_hook.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module ShopifyCLI
+  module Theme
+    module DevServer
+      class HotReload
+        class ReloadJSHook < Hook
+          def initialize
+
+          end
+
+          def call(mode)
+            # returns the correct JS output
+          end
+        end
+      end
+    end
+  end
+end
